Build psmcomm on *BSD systems only.
This is the second part of fixing the backend part of synaptics. The new build
system is now able to detect whether it's running on *BSD or on Linux and will
only build the backends that are available on these platforms. Also it will
remove any reference from the built driver to the non-available backends.
Thanks to Christian Schmitt for reporting.