ReadDevDimensions on opened device
The auto-dev probing requires opening the device locally to determine the type of device, so
that the right protocol can be assigned. However, all other setup work should be performed
during initialization of the assigned and opened device. This patch moves the
read-device-dimensions functionality from some special cases during the probe process to
the initialization routine, where it can be used for all supported devices.

Signed-off-by: Henrik Rydberg <rydberg@euromail.se>
Signed-off-by: Peter Hutterer <peter.hutterer@redhat.com>