Remove repeater functionality.
Remove the repeater functionality completely from the synaptics touchpad
driver. It is buggy in its current implementation and its usefulness is
questionable.

According to the INSTALL file, the repeater is there only for testing. In
fact, if a supported device is found even a configured repeater fifo is
automatically disabled. For most users the functionality is therefore
irrelevant and can be confusing. If I understand the workings of the
repeater correctly, a developer could instead just read the data directly
from an unsupported device's character special file under /dev
or /dev/input.

With today's more dynamic device configuration possibilities via udev and
hal it's also less likely that the synaptics driver would silently block
another devices data, this situation being for what the repeater
functionality seems to have been introduced in the past.

Signed-off-by: Peter Hutterer <peter.hutterer@redhat.com>