Add support for touchpads (such as Elantech) that do not report pressure
Change the driver to react to BTN_TOUCH if device does not report
ABS_PRESSURE since there are touchpads (such as Elantech) that do not
support pressure reading but otherwise are perfectly useable in absolute
mode.

Also fix error checking of ioctl calls (positive return value is not an
error, only negative is).

Signed-off-by: Dmitry Torokhov <dtor@mail.ru>
Signed-off-by: Peter Hutterer <peter.hutterer@redhat.com>