eventcomm: if we can't grab the device, assume it's not a touchpad.
If a device is configured in the xorg.conf, it will get added once in the
config, once through HAL. Since we grab the device, the second device will be
mute. And confuses tools and users alike.

So let's just assume that if we can't grab the device, it doesn't exist.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Christoph Brill <egore911@egore911.de>