Add model-specific edges for appletouch.
Needs around 8.5% to be useable.

I created a table to iterate through the different products rather than
using if/else branches.  I can enumerate the appletouch product_ids, but I
suspect they will all be around the same range (hence the PRODUCT_ANY).  If
another product id shows different behavior, we can just add an entry to
the table as appropriate.

I also changed the default eheight to be 5.4% to match the spec.

Reported-by: Jeremy Huddleston <jeremyhu@freedesktop.org>
Signed-off-by: Jeremy Huddleston <jeremyhu@freedesktop.org>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Henrik Rydberg <rydberg@euromail.se>