Auto-adjust edges if values fall outside queried min/max ranges. (#21001)
The kernel provides min/max for x/y values but still allows devices to send
coordinates outside this range. If the edges are autodetected, re-adjust the
edge settings to fit within the new effective min/max range.

When the edges change the property needs to be updated accordingly. This
can't be done immediately as changing properties requires mallocs and
HandleState is called during the signal handler.
Instead, set a timer to be called when the server isn't busy and update the
property then. The delay between setting the timer and sending the property
notify event also reduces the number of events sent, the property event
includes the latest state only.

If the edges were configured by the user, don't re-adjust.
This obsoletes the SpecialScrollAreaRight option as it provides the same
functionality, without the side-effects triggering 21001.

X.Org Bug 21001 <http://bugs.freedesktop.org/show_bug.cgi?id=21001>

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>