Add configurable x/y resolution to fix sensitivity on wide touchpads.
Synaptics uses anisotropic coordinate system.  On some wide touchpads
vertical resolution can be twice as high as horizontal which causes
unequal sensitivity on x/y directions.

VertResolution and HorizResolution can be used to set the resolution.
The ratio of the values is used to compensate x/y sensitivity.  The
properties are configured automatically if touchpad reports resolution
and if running on linux 2.6.31 or newer.

Fixes xorg bug #18351.

Signed-off-by: Tero Saarni <tero.saarni@gmail.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>