Split rc and ret into two different variables.
This is a cosmetic change only.

ioctls - though not the ones used here - may return non-zero non-negative
values to the caller. Using the return value as a boolean and as return
value for ioctls at the same time could result in false positives if one
ioctl returns 1. This cannot happen in this part of the code but a
separation of the two improves readability.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Christoph Brill <egore911@egore911.de>