Reshuffle to handle dead area better.
This patch makes input in the dead area essentially disappear to the client.
If a finger is detected outside the active area, it's coordinates and other
data is reset to zero. Edge detection and scrolling is skipped and the event
is not recorded in the movement history.

On touchpads that require this area, it is now possible to click the button
while leaving another finger on the touchpad.

Regression: moving into the dead area during a drag releases the drag button.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>