Add basic framework for unit-testing.
Also adds a bunch of fake symbols so we can link. Note that any of these
symbols will return false, 0 or whatever the zero value for the symbol is.
Care must be taken when writing test that this doesn't screw up the test.

Tests are always built but only run on make check.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>