eventcomm: move need_grab into a proto-specific struct.
Don't just assign a BOOL to proto_data, use a struct for readability.
This changes nothing in the code, especially since need_grab is always
false.

The current code assumes that if a proto_data field is present, we respect
the need_grab from this field. Otherwise, we always try to grab.
need_grab is however always FALSE, so all this is mostly for readability, if
not necessarily understandability...

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>