Revert "Default to 2-finger emulation when HW supports it"
This changes the default behavior for trackpads that have only pressure
information to emulate two finger actions. It's been reported that the
default value is too low and/or that the pressure values may fluctuate
with environmental factors (temperature, humidity, etc.). When the
value is wrong, spurious right clicks and scroll events are triggered.

Fixes: http://bugs.launchpad.net/bugs/742213

This reverts commit ffa6dc2809734a6aaa690e9133d6761480603a68.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>