Introduce POLL_MS for packet frequency
We expect to be receiving a steady 80 packets/sec (which gives 40
reports/sec with more than one finger on the pad, as Advanced Gesture
Mode requires two PS/2 packets per report).  Instead of a random
scattering of magic 13 and 20ms numbers scattered throughout the driver,
introduce POLL_MS as 14ms.

Having this here allows us to call back at a steady frequency to ensure
that the finger motion remains steady.

Signed-off-by: Daniel Stone <daniel@fooishbar.org>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>