Revert "build: apply the distcheck tricks used in xf86-input-evdev"
This reverts commit 6eb829e07e455a8a04dabae7f257dd42a9b8bcdf.

Althought this feature is desirable, it introduces a dormant bug.

The value of sdkdir during distcheck points to an empty xserver sdkdir
where xserver-properties.h is expected by the tools.

Specifying XORG_CFLAGS for the tools worked around the issue by supplying
the real path to the xserver sdk. tools must not depend on xserver.
In any case it would be a brittle workaround.

--with-xorg-conf-dir='$${sysconfdir}/xorg':
althought this value will work, it brings some confusion as it is
unrelated to the where the configuration files are normally placed,
relative to $prefix which is /usr/share/X11 and not /etc.

This feature can be reintroduced later with the proper solution.

Signed-off-by: Gaetan Nadon <memsize@videotron.ca>