Revert "build: install documentation as part of make install."
This reverts commit d27b4e560ccf61b94f067156c6d5c1e3d3e5e1eb.

ChangeLog is generated from git and should be created at dist time only.
The original patch creates it at make time multiple times and fails
when user permissions change such as when using sudo.

Some have expressed a desire to install files such as ChangeLog and README.
This reversal does not dismiss the idea, but simply reverts a non-essential
change that happens to have an implementation bug just before making a module
release.

Anyone is welcome to promote the idea in the context of the X.Org project
where all modules would exhibit the same behaviour. The current behaviour of
ChangeLog, although not perfect, has been thoroughly reviewed and widely
accepted for several years.

As for the files in docs, they should be re-submitted for reviews to determine
if they constitute users docs that should be installed as opposed to
developers docs which are not.

Reported-by: Daniel Stone <daniel@fooishbar.org>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>