Revert "Replace the motion estimator"
The algorithm is completely wrong because it uses subtraction of
unsigned variables. The negative effects of this commit manifest in
cursor warps to edges or corners of the screen.

Since the algorithm has never worked right, previous testing must be
disregarded. Revert it until we have a tested algorithm.

This reverts commit b26125e412a130b7a8f8b6adf9ffc8e9cc8df42c.

Conflicts:

	src/synaptics.c

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>