Ignore motion during touch count changes on semi-mt devices
Semi-mt devices do not track touches. The locations of touches are
unknown, we only have the bounding box of two of them. When the number of
fingers changes, the bounding box coordinates may change as well, but
the cumulative relative motion updates at that instant are invalid.

To work around this, ignore changes in cumulative relative motion if the
touch count changes.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>