Include open but unchanged touches when guessing clickfingers
On a clickpad, when the button is pressed the clickfinger guessing will
only work properly if all touches have new data to report. If a touch
has not changed, then it will not be counted. This leads to inaccurate
finger counts.

This change ensures that all active touches are counted. Note that the X
and Y valuators of active but unchanged touches are still valid.

Signed-off-by: Chase Douglas <chase.douglas@canonical.com>
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>