Fix coasting speed trigger
CoastingSpeed is defined as scrolls/s. The previous code just used
delta/seconds which depended on the device coordinate range and exceeded the
default CoastingSpeed at almost any scroll event.

Divide the estimated delta by the scroll distance to get the accurate
scrolls/s number. Since that now changes the contents of what's in
coast_speed_y, change the users of that too.

http://bugzilla.redhat.com/813686

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>